Lesson 5

+++ src/main/java/ru/armishev/intellekta/jpa/SalesPeriodJpaRepository.java
@@ -21,7 +21,7 @@ public interface SalesPeriodJpaRepository extends JpaRepository<SalesPeriod, Lon
     List<SalesPeriod> findByProductName(String productName);
 
     @Query(value = "SELECT MAX(id) from sales_period", nativeQuery = true)
-    Integer getMaxId();
+    long getMaxId();
 
     List<SalesPeriod> findByPrice(long price);
